AJAX Nav: implemented new-style hijacking
Before this commit, the hijacking was accomplished when the page
was loaded, by a snippet to be put in the very beginning of the
HEAD element.

This approach has a number of issues, because it makes the browser
to render stuff that is then unused; moreover it causes the double
back problem when landing from a referring site.

The new approach instead doesn't hijack incoming requests, rather
it hijacks clicks on "a" elements, by inspecting a list of hijackable
paths previously defined by the developer.

Documentation has been updated, check it out for more information.