Demo: fixed issue #2 reported by @jmckitrick
The whole point is: the demo is not fully functional when run via
bare HTML files. You need a backend that returns different response
when the request comes from XMLHttpRequest or it is a full page load.

This commit makes the backend return the contents of the "index-xhr.html"
file when the root is requested via XHR. :-)